% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icu-package.R
\docType{package}
\name{icu-package}
\alias{icu}
\alias{icu-package}
\title{icu: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/icu}
  \item Report bugs at \url{https://github.com/DavisVaughan/icu/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
