% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrappers.R
\name{icu}
\alias{get_native_enc}
\alias{icu}
\alias{stri_length}
\title{Basic string processing}
\usage{
get_native_enc()

stri_length(x)
}
\arguments{
\item{x}{a character vector}
}
\description{
Functions to manage and manipulate unicode strings.
}
\examples{
# Show native encoding
icu::get_native_enc()
utils::localeToCharset()
# Count length of strings
x <- c("Paris", "Zürich", "\\346\\235\\261\\344\\272\\254")
icu::stri_length(x)
base::nchar(x)
}

