% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{clean_names}
\alias{glue}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{glue}{\code{\link[glue:glue]{glue()}}}

  \item{janitor}{\code{\link[janitor:clean_names]{clean_names()}}}
}}

